common/requirements.txt
PyPDF2
google-generativeai
python-dotenv
psycopg2-binary
boto3
python-docx